Refactor observables
